+++ docs/changelog.md
@@ -6,6 +6,7 @@ __Node.py__
 - Fix #48 -- Pip Install on Windows: Underlying buffer has been detached
 - Fix #49 -- Importing local Python packages should be preferred over globally installed packages
 - `nodepy.print_exc()` now tolerates `sys.stderr` not having an `isatty()` member
+- Support `import * from 'module-name'` syntax with `!require-import-syntax` extension
 
 __nppm__